Renamed manage_logs to log_manager
